+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 ## 2026-06-19
 - Audit globeview (thread atlas)
+- Sphere UI upgrades